\name{CFL-package}
\alias{CFL-package}
\alias{CFL}
\docType{package}
\title{
  Compensatory Fuzzy Logic
}
\description{
  The main results on this project are: a package to analyze the data based on compensatory fuzzy logic algorithms and a reasoner for the compensatory fuzzy logic.
}
\details{
\tabular{ll}{
Package: \tab CFL\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-08-17\cr
License: \tab Unlimited\cr
LazyLoad: \tab yes\cr
}
}
\author{
Pablo Michel Marin Ortega, Kornelius Rohmeyer

Maintainer: Pablo Michel Marin Ortega <pablomo@uclv.edu.cu>

}
\references{
Espin, R., Fern\'{a}ndez, E., Mazcorro, G., Marx-G\'{o}mez J., M.I. Lecich:
Compensatory Logic: A fuzzy normative model for decision making. Investigaci\'{o}n
Operativa. Universidad de la Habana.Vol. 27, 2, pp. 188-197 (2006).

Espin, R., Mazcorro, G., Fen\'{a}ndez, E.: Consideraciones sobre el car\'{a}cter
normativo de la l\'{o}gica difusa compensatoria. Infraestructura de Datos
Espaciales en Iberoam\'{e}rica y el Caribe. IDICT. Cuba (2007).

Espin, R., Fern\'{a}ndez, E.: La L\'{o}gica Difusa Compensatoria: Una Plataforma
para el Razonamiento y la Representaci\'{o}n del Conocimiento en un Ambiente
de Decisi\'{o}n Multicriterio. In: An\'{a}lisis Multicriterio para la Toma
de Decisiones: M\'{e}todos y Aplicaciones. Coedici\'{o}n: editorial Plaza
y Valdes/editorial Universidad de Occidente (2009).
}
\keyword{ package }
\examples{
ldc.conj(0,1,0,1)
}
